Refactor: Improve settings page and theme context
This commit includes several changes to improve the settings page and theme context:

- Refactor: Move settings items to separate components
- Feat: Add settings section component
- Fix: Use settings store directly in GeminiApiKey component
- Fix: Use settings store directly in ThemeContext
- Fix: Update path aliases in tsconfig and vite config
- Fix: Update useGoogleDriveSync useEffect dependency array
