+++ src/store/settings.ts
@@ -1,5 +1,6 @@
-import { DBStore } from '../db';
-import { StoreNames } from '../types/db.types';
+// filepath: y:\All-Projects\Study-Pal\src\store\settings.ts
+import { DBStore } from '@db';
+import { StoreNames } from '@type/db.types';
 
 /**
  * Setting keys - enum of all available settings
@@ -8,7 +9,6 @@ import { StoreNames } from '../types/db.types';
 export enum SettingKeys {
   ACTIVE_THEME = 'active_theme',
   GEMINI_API_KEY = 'gemini_api_key',
-  // Add more settings as needed:
 }
 
 /**
@@ -16,112 +16,36 @@ export enum SettingKeys {
  * Extends the generic DBStore to provide type-safe access to settings
  */
 class SettingsStore extends DBStore<any> {
-  constructor() {
-    super(StoreNames.SETTINGS);
-  }
+    constructor() {
+        super(StoreNames.SETTINGS);
+    }
 
-  /**
-   * Get the currently selected theme name
-   */
-  async getTheme(): Promise<string | undefined> {
-    return this.get(SettingKeys.ACTIVE_THEME);
-  }
+    // Active Theme Setting
+    get activeTheme(): Promise<string | undefined> {
+        return this.get(SettingKeys.ACTIVE_THEME);
+    }
 
-  /**
-   * Set the theme name
-   */
-  async setTheme(themeId: string): Promise<void> {
-    await this.set(SettingKeys.ACTIVE_THEME, themeId);
-  }
-  
-  /**
-   * Get the Gemini API key
-   */
-  async getGeminiApiKey(): Promise<string | undefined> {
-    return this.get(SettingKeys.GEMINI_API_KEY);
-  }
-  
-  /**
-   * Set the Gemini API key
-   */
-  async setGeminiApiKey(apiKey: string): Promise<void> {
-    await this.set(SettingKeys.GEMINI_API_KEY, apiKey);
-  }
+    set activeTheme(value: string) {
+        this.set(SettingKeys.ACTIVE_THEME, value);
+    }
 
-  /**
-   * Get a setting by key with strong typing
-   */
-  async getSetting<T>(key: SettingKeys): Promise<T | undefined> {
-    return this.get(key) as Promise<T | undefined>;
-  }
+    // Gemini API Key Setting
+    get geminiApiKey(): Promise<string | undefined> {
+        return this.get(SettingKeys.GEMINI_API_KEY);
+    }
 
-  /**
-   * Set a setting with strong typing
-   */
-  async setSetting<T>(key: SettingKeys, value: T): Promise<void> {
-    await this.set(key, value);
-  }
+    set geminiApiKey(value: string) {
+        this.set(SettingKeys.GEMINI_API_KEY, value);
+    }
 
-  /**
-   * Reset all settings to their defaults
-   */
-  async resetAll(): Promise<void> {
-    await this.clear();
-  }
+    /**
+     * Reset all settings to their defaults
+     */
+    async resetAll(): Promise<void> {
+        await this.clear();
+    }
 }
 
 // Create and export a singleton instance
-export const settingsStore = new SettingsStore();
-
-/**
- * Retrieves the saved theme name
- * @returns A promise that resolves with the theme name or undefined
- */
-export async function getThemeSetting(): Promise<string | undefined> {
-  return settingsStore.getTheme();
-}
-
-/**
- * Saves the selected theme name
- * @param themeName - The name of the theme to save
- * @returns A promise that resolves when the theme name is saved
- */
-export async function setThemeSetting(themeName: string): Promise<void> {
-  await settingsStore.setTheme(themeName);
-}
-
-/**
- * Retrieves the saved Gemini API key
- * @returns A promise that resolves with the API key or undefined
- */
-export async function getGeminiApiKey(): Promise<string | undefined> {
-  return settingsStore.getGeminiApiKey();
-}
-
-/**
- * Saves the Gemini API key
- * @param apiKey - The API key to save
- * @returns A promise that resolves when the API key is saved
- */
-export async function setGeminiApiKey(apiKey: string): Promise<void> {
-  await settingsStore.setGeminiApiKey(apiKey);
-}
-
-// Export legacy functions to maintain backward compatibility
-// So existing code doesn't need to be updated all at once
-export async function getSettings(): Promise<Record<string, any>> {
-  const keys = await settingsStore.getAllKeys();
-  const result: Record<string, any> = {};
-  
-  for (const key of keys) {
-    result[key] = await settingsStore.get(key);
-  }
-  
-  return result;
-}
-
-export async function updateSettings(partialSettings: Record<string, any>): Promise<void> {
-  for (const [key, value] of Object.entries(partialSettings)) {
-    await settingsStore.set(key, value);
-  }
-}
+const settingsStore = new SettingsStore();
+export default settingsStore;
