+++ tsconfig.app.json
@@ -32,6 +32,7 @@
       "@contexts/*": ["contexts/*"],
       "@hooks/*": ["hooks/*"],
       "@store/*": ["store/*"],
+      "@type/*": ["types/*"],
       "@utils/*": ["utils/*"],
       "@db": ["db.ts"],
     }
